Move and rename locust.inspectlocust to locust.user.inspectuser
